+++ packages/liveblocks-core/src/globals/augmentation.ts
@@ -76,4 +76,15 @@ export type DE = GetOverride<"RoomEvent", Json, "is not a valid JSON value">;
 export type DM = GetOverride<"ThreadMetadata", BaseMetadata>;
 
 export type DRI = GetOverride<"RoomInfo", BaseRoomInfo>;
-export type DAD = GetOverride<"ActivitiesData", BaseActivitiesData>;
+export type DAD = GetOverrideOrErrorValue<
+  "ActivitiesData",
+  BaseActivitiesData,
+  {
+    [K in keyof Liveblocks["ActivitiesData"]]: `At least one of the custom notification kinds you provided for 'ActivitiesData' does not match its requirements. To learn how to fix this, see https://liveblocks.io/docs/errors/ActivitiesData`;
+  }
+>;
+
+export type KDAD = keyof DAD extends `$${string}`
+  ? keyof DAD
+  : `Your custom 'ActivitiesData' type can only contain kind names that start with '$'. To learn how to fix this, see https://liveblocks.io/docs/errors/ActivitiesData`;
+// XXX Change error message to include link